LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/radio-focus-navigation-group-focusable-focus-expected.txt


enabled 1  Button  enabled 2  enabled 3


FAIL Focusable elements between radio elements in a group can be focused. assert_equals: expected Element node <input type="radio" name="radio" id="a"></input> but got Element node <input type="text" id="end"></input>
FAIL During arrow key navigation, focus and checked is set on the radio element. assert_equals: expected Element node <input type="radio" name="radio" id="a"></input> but got Element node <input type="text" id="end"></input>
